v10 Release Candidate I
Well, I think we're very close until devSin arrives and tells me I've
done it all wrong. :)

Actual changes are pretty minimal--generally the changes discussed with
Hurricane in the effing effs thread and a few last changes from a BGEE
review. Internal docs are updated to this build; though I still need to
run a spellcheck on both the docs and readme. Also need to test this on
a SoA-only install.

Translators will be unhappy, as we ended up stuffing in a lot of new
strings after we said it was 'finalized' a couple of months ago.